+++ src/models/orders/ordersSchema.js
@@ -3,9 +3,8 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
     return (mod && mod.__esModule) ? mod : { "default": mod };
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.addOrderValidation = exports.Orders = void 0;
+exports.Orders = void 0;
 const mongoose_1 = __importDefault(require("mongoose"));
-const joi_1 = __importDefault(require("joi"));
 const { Schema, model } = mongoose_1.default;
 const orderSchema = new Schema({
     order: [
@@ -51,16 +50,4 @@ orderSchema.pre("save", (error, next) => {
     next();
 });
 exports.Orders = model("Orders", orderSchema);
-exports.addOrderValidation = joi_1.default.object({
-    order: joi_1.default.array().items(joi_1.default.object({
-        dishName: joi_1.default.string().required(),
-        image: joi_1.default.string().required(),
-        price: joi_1.default.string().required(),
-        number: joi_1.default.number().required(),
-        shop: joi_1.default.string().required(),
-    })),
-    totalPrice: joi_1.default.string().required(),
-    dateOrder: joi_1.default.string().required(),
-    owner: joi_1.default.any(),
-});
 //# sourceMappingURL=ordersSchema.js.map
