+++ src/validation-schemas.ts/commons.js
@@ -1,7 +1,9 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.loginUserSchema = exports.registerUserSchema = void 0;
+exports.addOrderValidation = exports.loginUserSchema = exports.registerUserSchema = void 0;
 var user_travel_validation_1 = require("./user-travel-validation");
 Object.defineProperty(exports, "registerUserSchema", { enumerable: true, get: function () { return user_travel_validation_1.registerUserSchema; } });
 Object.defineProperty(exports, "loginUserSchema", { enumerable: true, get: function () { return user_travel_validation_1.loginUserSchema; } });
+var orders_validation_1 = require("./orders-validation");
+Object.defineProperty(exports, "addOrderValidation", { enumerable: true, get: function () { return orders_validation_1.addOrderValidation; } });
 //# sourceMappingURL=commons.js.map
