+++ src/validation-schemas.ts/shops-validation.js
@@ -1,0 +1,2 @@
+"use strict";
+//# sourceMappingURL=shops-validation.js.map
