+++ typeScript/models/user/userSchema.ts
@@ -1,5 +1,4 @@
 import mongoose from "mongoose";
-import Joi from "joi";
 import express from 'express';
 import { IUserDelivery } from "../../types/commons";
 
@@ -32,10 +31,3 @@ userSchema.post("save", (error: any, next: express.NextFunction) => {
 });
 
 export const Users = model<IUserDelivery>("Users", userSchema);
-
-export const addUserValidation = Joi.object({
-  name: Joi.string().max(30).required(),
-  email: Joi.string().email().required(),
-  phone: Joi.string().required(),
-  address: Joi.string().required(),
-});
