+++ typeScript/validation-schemas/user-delivery-validation.ts
@@ -1,0 +1,8 @@
+import Joi from "joi";
+
+export const addUserDeliveryValidation = Joi.object({
+  name: Joi.string().max(30).required(),
+  email: Joi.string().email().required(),
+  phone: Joi.string().required(),
+  address: Joi.string().required(),
+});
